requirements.txt
streamlit
sqlalchemy
pandas
plotly-express